Move facings logic to UnitFacings class
- for #63
