+++ src/main/java/com/fundynamic/d2tm/game/entities/units/UnitFacings.java
@@ -1,6 +1,8 @@
 package com.fundynamic.d2tm.game.entities.units;
 
 
+import com.fundynamic.d2tm.math.Vector2D;
+
 import java.util.HashMap;
 import java.util.Map;
 
@@ -35,4 +37,23 @@ public enum UnitFacings {
     public static UnitFacings byId(int id) {
         return facingsById.get(id);
     }
+
+    public static UnitFacings determine(Vector2D from, Vector2D to) {
+        boolean left = to.getXAsInt() < from.getXAsInt();
+        boolean right = to.getXAsInt() > from.getXAsInt();
+        boolean up = to.getYAsInt() < from.getYAsInt();
+        boolean down = to.getYAsInt() > from.getYAsInt();
+
+        if (up && left) return UnitFacings.LEFT_UP;
+        if (up && right) return UnitFacings.RIGHT_UP;
+        if (down && left) return UnitFacings.LEFT_DOWN;
+        if (down && right) return UnitFacings.RIGHT_DOWN;
+        if (up) return UnitFacings.UP;
+        if (down) return UnitFacings.DOWN;
+        if (left) return UnitFacings.LEFT;
+        if (right) return UnitFacings.RIGHT;
+
+        return UnitFacings.RIGHT;
+    }
+
 }
